requirements.txt
polib==1.0.7
PyYAML==3.11
